Refactor controls management
