+++ packages/vue-mapbox-gl/composables/index.js
@@ -1,3 +1,4 @@
-export * from './useMap.js';
+export * from './useControl.js';
 export * from './useEventsBinding.js';
+export * from './useMap.js';
 export * from './usePropsBinding.js';
